add $wgSecurePollEditOtherWikis
Add a configuration variable that can be set to false if this wiki
should only create local elections, and can be set to true if this
wiki is a central votewiki such as vote.wikimedia.org.

The InitializeSettings.php patch should be merged and deployed
first, so that vote.wikimedia.org doesn't break.

Bug: T384302
Depends-On: I22c48a976b36b740d72a152eb53d39ddc4c4748d
Change-Id: Ifd7d8f125306b8fd17118867cefb5dd19806d049